Create final_project_week7.py
Project Title:
Log Analysis Using Regular Expressions

Introduction
Imagine your company uses a server that runs a service called ticky, an internal ticketing system. The service logs events to syslog, both when it runs successfully and when it encounters errors.

The service's developers need your help getting some information from those logs so that they can better understand how their software is used and how to improve it. So, for this lab, you'll write some automation scripts that will process the system log and generate reports based on the information extracted from the log files.

What you'll do
Use regex to parse a log file
Append and modify values in a dictionary
Write to a file in CSV format
Move files to the appropriate directory for use with the CSV->HTML converter
You'll have 90 minutes to complete this lab.